Speech_to_text/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
torch